emotion_manga/requirements.txt
torch
diffusers
transformers
Pillow
deepface
opencv-python
gradio
matplotlib
